fix: check result of running command
When insights-client is not installed or it is not possible
to run insights-client for other reason (wrong permissions,
SELinux, etc.), then running command is finished with
error and it is not possible to check state of the process.
When it is not possible to run insights-client, then print
error in the output of rhc.